Seperated out beeper.cpp code
