+++ basic/test.bsc
@@ -1,4 +1,4 @@
-cat "x":end
+sfx 0,8:end
 
 a = 11:b$ = "$12"c = 13:d = 14
 a1 = 43
